Create client interface
