+++ cmd/topicctl/subcmd/get.go
@@ -87,7 +87,7 @@ func getRun(cmd *cobra.Command, args []string) error {
 	ctx := context.Background()
 	sess := session.Must(session.NewSession())
 
-	var adminClient *admin.Client
+	var adminClient admin.Client
 	var clientErr error
 
 	if getConfig.clusterConfig != "" {
@@ -97,9 +97,9 @@ func getRun(cmd *cobra.Command, args []string) error {
 		}
 		adminClient, clientErr = clusterConfig.NewAdminClient(ctx, sess, true)
 	} else {
-		adminClient, clientErr = admin.NewClient(
+		adminClient, clientErr = admin.NewZKAdminClient(
 			ctx,
-			admin.ClientConfig{
+			admin.ZKAdminClientConfig{
 				ZKAddrs:  []string{getConfig.zkAddr},
 				ZKPrefix: getConfig.zkPrefix,
 				Sess:     sess,
